Remove exception-raising condition in ::after
This was checking that the specified method names were in fact instance methods and raising an exception if any were not. In practice, ::after blocks will be at the top of an including module, in which case no instance methods will have been defined. Relaxing the constraint in light of this.